add integration test for invalid format and non-existent revision description with some minor formatting fixes
